+++ domain/src/commonMain/kotlin/io/github/alelk/pws/domain/core/result/DeleteResourceResult.kt
@@ -13,3 +13,19 @@ sealed interface DeleteResourceResult<out R : Any> {
     val message: String = exception?.message ?: exception?.let { "Unknown error: ${it::class.simpleName}" } ?: "Unknown error"
   ) : DeleteResourceResult<R>
 }
+
+fun <A : Any, B : Any> DeleteResourceResult<A>.map(transform: (A) -> B): DeleteResourceResult<B> =
+  when (this) {
+    is DeleteResourceResult.Success -> DeleteResourceResult.Success(resource = transform(this.resource))
+    is DeleteResourceResult.NotFound -> DeleteResourceResult.NotFound(resource = transform(this.resource))
+    is DeleteResourceResult.ValidationError -> DeleteResourceResult.ValidationError(
+      resource = transform(this.resource),
+      message = this.message
+    )
+
+    is DeleteResourceResult.UnknownError -> DeleteResourceResult.UnknownError(
+      resource = transform(this.resource),
+      exception = this.exception,
+      message = this.message
+    )
+  }