+++ domain/src/commonMain/kotlin/io/github/alelk/pws/domain/favorite/model/Favorite.kt
@@ -1,6 +1,5 @@
 package io.github.alelk.pws.domain.favorite.model
 
-import io.github.alelk.pws.domain.core.ids.SongNumberId
 import kotlin.time.ExperimentalTime
 import kotlin.time.Instant
 
@@ -9,7 +8,7 @@ import kotlin.time.Instant
  */
 @OptIn(ExperimentalTime::class)
 data class Favorite(
-  val songNumberId: SongNumberId,
+  val subject: FavoriteSubject,
   val addedAt: Instant
 )