+++ domain/src/commonMain/kotlin/io/github/alelk/pws/domain/favorite/usecase/AddFavoriteUseCase.kt
@@ -1,18 +1,18 @@
 package io.github.alelk.pws.domain.favorite.usecase
 
-import io.github.alelk.pws.domain.core.result.ToggleResourceResult
+import io.github.alelk.pws.domain.core.result.CreateResourceResult
 import io.github.alelk.pws.domain.core.transaction.TransactionRunner
+import io.github.alelk.pws.domain.favorite.model.Favorite
 import io.github.alelk.pws.domain.favorite.model.FavoriteSubject
 import io.github.alelk.pws.domain.favorite.repository.FavoriteWriteRepository
 
 /**
- * Use case: toggle favorite status for a song.
+ * Use case: add song to favorites.
  */
-class ToggleFavoriteUseCase(
+class AddFavoriteUseCase(
   private val favoriteRepository: FavoriteWriteRepository,
   private val txRunner: TransactionRunner
 ) {
-  suspend operator fun invoke(subject: FavoriteSubject): ToggleResourceResult<FavoriteSubject> =
-    txRunner.inRwTransaction { favoriteRepository.toggle(subject) }
+  suspend operator fun invoke(subject: FavoriteSubject): CreateResourceResult<Favorite> =
+    txRunner.inRwTransaction { favoriteRepository.add(subject) }
 }
-