+++ domain/src/commonMain/kotlin/io/github/alelk/pws/domain/favorite/usecase/RemoveFavoriteUseCase.kt
@@ -1,7 +1,8 @@
 package io.github.alelk.pws.domain.favorite.usecase
 
-import io.github.alelk.pws.domain.core.ids.SongNumberId
+import io.github.alelk.pws.domain.core.result.DeleteResourceResult
 import io.github.alelk.pws.domain.core.transaction.TransactionRunner
+import io.github.alelk.pws.domain.favorite.model.FavoriteSubject
 import io.github.alelk.pws.domain.favorite.repository.FavoriteWriteRepository
 
 /**
@@ -11,7 +12,7 @@ class RemoveFavoriteUseCase(
   private val favoriteRepository: FavoriteWriteRepository,
   private val txRunner: TransactionRunner
 ) {
-  suspend operator fun invoke(songNumberId: SongNumberId): Boolean =
-    txRunner.inRwTransaction { favoriteRepository.remove(songNumberId) }
+  suspend operator fun invoke(subject: FavoriteSubject): DeleteResourceResult<FavoriteSubject> =
+    txRunner.inRwTransaction { favoriteRepository.remove(subject) }
 }